+++ include/mesh_to_vector.h
@@ -1,9 +1,11 @@
 #ifndef MESH_TO_VECTOR_H
 #define MESH_TO_VECTOR_H
 
+#include <stdlib.h>
+
 #include "mesh.h"
 #include "vector.h"
 
-void mesh_to_vector(const Mesh *mesh, int index, Vector *v);
+void mesh_to_vector(const Mesh *mesh, size_t index, Vector *v);
 
 #endif
